requirements.txt
flask
pyspellchecker
pycld3
symspellpy
